# backend_go/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.referralSettingsPayload:
    properties:
      referral_percentage:
        type: number
      referral_program_enabled:
        type: boolean
    type: object
  handlers.registerBotUserPayload:
    properties:
      telegram_id:
        type: integer
    type: object
...
  handlers.updateUserCaptchaStatusPayload:
    properties:
      has_passed_captcha:
        type: boolean
    type: object
  models.BotUser:
    properties:
      hasPassedCaptcha:
        type: boolean
      id:
        type: integer
      isDeleted:
        type: boolean
      telegramID:
        type: integer
    type: object
  models.BotUserResponse:
    properties:
      balance:
        type: number
      has_passed_captcha:
        type: boolean
      id:
        type: integer
      is_deleted:
        type: boolean
      telegram_id:
        type: integer
    type: object
  models.Category:
    properties:
      id:
        type: integer
      name:
        type: string
      parent:
        $ref: '#/definitions/models.Category'
      parentID:
        description: Указатель, чтобы разрешить NULL
        type: integer
      products:
        items:
          $ref: '#/definitions/models.Product'
        type: array
      subCategories:
        items:
          $ref: '#/definitions/models.Category'
        type: array
    type: object
...
  models.Order:
    properties:
      Product:
        $ref: '#/definitions/models.Product'
      amount:
        type: number
      created_at:
        type: string
      id:
        type: integer
      product_id:
        type: integer
      quantity:
        type: integer
      status:
        type: string
      user_id:
        type: integer
    type: object
...
  models.Product:
    properties:
      category:
        $ref: '#/definitions/models.Category'
      category_id:
        type: integer
      details:
        type: string
      id:
        type: integer
      name:
        type: string
      price:
        type: number
      subscription_period_days:
        type: integer
      type:
        type: string
    type: object
...
  models.TransactionResponse:
    properties:
      amount:
        type: number
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      order_id:
        type: integer
      type:
        $ref: '#/definitions/models.TransactionType'
      user_id:
        type: integer
    type: object
  models.TransactionType:
    enum:
    - deposit
    - purchase
    type: string
    x-enum-varnames:
    - Deposit
    - Purchase
  models.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      referral_percentage:
        type: number
      referral_program_enabled:
        type: boolean
      role:
        $ref: '#/definitions/models.UserRole'
    type: object
  models.UserRole:
    enum:
    - admin
    - seller
    type: string
    x-enum-varnames:
    - Admin
    - Seller
  models.UserSubscription:
    properties:
      Product:
        $ref: '#/definitions/models.Product'
      bot_user:
        $ref: '#/definitions/models.BotUser'
      bot_user_id:
        type: integer
      created_at:
        type: string
      details:
        type: object
      expires_at:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      order:
        $ref: '#/definitions/models.Order'
      order_id:
        type: integer
      product_id:
        type: integer
      provisioned_id:
        type: string
      updated_at:
        type: string
    type: object
  responses.BalanceResponse:
    properties:
      balance:
        type: number
    type: object
  responses.ErrorResponseSchema:
    description: This is the standard error response structure.
    properties:
      data: {}
      error:
        type: string
      success:
        example: false
        type: boolean
    type: object
  responses.MessageResponse:
    properties:
      message:
        type: string
    type: object
  responses.RegisterBotUserResponse:
    properties:
      has_passed_captcha:
        type: boolean
      is_new:
        type: boolean
      user:
        $ref: '#/definitions/models.BotUserResponse'
    type: object
...
  responses.ResponseSchema-array_models_Order:
    properties:
      data:
        items:
          $ref: '#/definitions/models.Order'
        type: array
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
  responses.ResponseSchema-array_models_TransactionResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/models.TransactionResponse'
        type: array
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
  responses.ResponseSchema-array_models_UserSubscription:
    properties:
      data:
        items:
          $ref: '#/definitions/models.UserSubscription'
        type: array
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
  responses.ResponseSchema-models_BotUserResponse:
    properties:
      data:
        $ref: '#/definitions/models.BotUserResponse'
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
  responses.ResponseSchema-models_UserResponse:
    properties:
      data:
        $ref: '#/definitions/models.UserResponse'
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
  responses.ResponseSchema-responses_BalanceResponse:
    properties:
      data:
        $ref: '#/definitions/responses.BalanceResponse'
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
  responses.ResponseSchema-responses_MessageResponse:
    properties:
      data:
        $ref: '#/definitions/responses.MessageResponse'
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
  responses.ResponseSchema-responses_RegisterBotUserResponse:
    properties:
      data:
        $ref: '#/definitions/responses.RegisterBotUserResponse'
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
  responses.ResponseSchema-responses_SellerSettingsResponse:
    properties:
      data:
        $ref: '#/definitions/responses.SellerSettingsResponse'
      error:
        example: "null"
        type: string
        x-nullable: "true"
      success:
        example: true
        type: boolean
    type: object
...
  responses.SellerSettingsResponse:
    properties:
      id:
        type: integer
      referral_percentage:
        type: number
      referral_program_enabled:
        type: boolean
    type: object
...
host: localhost:8000
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API for your project.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Your Project API
  version: "1.0"
paths:
...
  /me:
    get:
      description: Retrieves details for the currently authenticated admin/seller
        user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-models_UserResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: Get Current User
      tags:
      - Users
  /me/referral-settings:
    put:
      consumes:
      - application/json
      description: Updates the referral program settings for the current admin/seller.
      parameters:
      - description: Referral settings
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/handlers.referralSettingsPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-responses_MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ApiKeyAuth: []
      summary: Update Referral Settings
      tags:
      - Users
      - Referrals
...
  /users/{telegram_id}:
    get:
      description: Retrieves bot user details by their Telegram ID.
      parameters:
      - description: User Telegram ID
        in: path
        name: telegram_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-models_BotUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ServiceApiKeyAuth: []
      summary: Get Bot User by Telegram ID
      tags:
      - Users
  /users/{telegram_id}/balance:
    get:
      description: Retrieves the current balance for a bot user.
      parameters:
      - description: User Telegram ID
        in: path
        name: telegram_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-responses_BalanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ServiceApiKeyAuth: []
      summary: Get User Balance
      tags:
      - Users
      - Balance
  /users/{telegram_id}/captcha-status:
    put:
      consumes:
      - application/json
      description: Updates the captcha verification status for a bot user.
      parameters:
      - description: User Telegram ID
        in: path
        name: telegram_id
        required: true
        type: integer
      - description: Captcha status
        in: body
        name: status
        required: true
        schema:
          $ref: '#/definitions/handlers.updateUserCaptchaStatusPayload'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-responses_MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ServiceApiKeyAuth: []
      summary: Update User Captcha Status
      tags:
      - Users
  /users/{telegram_id}/orders:
    get:
      description: Retrieves the order history for a bot user.
      parameters:
      - description: User Telegram ID
        in: path
        name: telegram_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-array_models_Order'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ServiceApiKeyAuth: []
      summary: Get User Orders
      tags:
      - Users
      - Orders
  /users/{telegram_id}/subscriptions:
    get:
      description: Retrieves a list of a bot user's active and expired subscriptions.
      parameters:
      - description: User Telegram ID
        in: path
        name: telegram_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-array_models_UserSubscription'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ServiceApiKeyAuth: []
      summary: Get User Subscriptions
      tags:
      - Users
      - Subscriptions
  /users/{telegram_id}/transactions:
    get:
      description: Retrieves the transaction history for a bot user.
      parameters:
      - description: User Telegram ID
        in: path
        name: telegram_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-array_models_TransactionResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ServiceApiKeyAuth: []
      summary: Get User Transactions
      tags:
      - Users
      - Transactions
  /users/register:
    post:
      consumes:
      - application/json
      description: Registers a new user from a Telegram bot or reactivates a deleted
        one.
      parameters:
      - description: User Telegram ID
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.registerBotUserPayload'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/responses.ResponseSchema-responses_RegisterBotUserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      security:
      - ServiceApiKeyAuth: []
      summary: Register a Bot User
      tags:
      - Users
  /users/seller-settings:
    get:
      description: Retrieves public settings for the seller, like referral program
        status.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/responses.ResponseSchema-responses_SellerSettingsResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/responses.ErrorResponseSchema'
      summary: Get Seller Settings
      tags:
      - Users
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  ServiceApiKeyAuth:
    in: header
    name: X-API-KEY
    type: apiKey
swagger: "2.0"